Added tests for invalid state, no auth-code in callback Moving the stubs setup to mock-server Tests only concerned with assertions now
