+++ integration-tests/oauth-mock-test-runner/src/main/groovy/com/okta/test/mock/wiremock/TestUtils.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.okta.test.mock.tests
+package com.okta.test.mock.wiremock
 
 class TestUtils {
     static byte[] toIntegerBytes(final BigInteger bigInt) {
